Main code.
This is just a place to store and review the code. I know that it is
horrendously inefficient, I know that using something like cuda would
speed it up ridiculously: It was taking 10 mins to render with a 950
and 12 GB of RAM when I set it to 1000 iterations. I am going to
rewrite this in js and make a meteor app with it in the coming weeks so
this should be merely a fossil. Gladly departed fossil.